Rework screen text to be more optimized
It is now split into multiple types of screen text: static, variable, and if.

Static screen text behaves almost like the previous implementation. It displays a static text, but now it is able to be reused and is poolable.

Variable screen text have a callable variable instead of static text. This callable will return what to display each time it is rendered. It is also support formatting out of the box which can edit attributes such as color, boldness, italics, etc.

The final screen text is 'if' screen text, it is essentially a wrapper around a callable function which returns the real screen text. This allows screen text to be more dynamic such as if-ing a variable and return different screen text depending on the answer.

In total this means there is no longer a need to create lots of ScreenText  instances each frame, all around improving performance and reusability.